small_backend/requirements.txt
aiofiles==23.2.1
aiogram==3.13.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==43.0.1
diff-match-patch==20230430
Django==5.0
django-autocomplete-light==3.11.0
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-filter==24.3
django-guardian==2.4.0
django-import-export==4.1.1
django-keycloak-auth==1.0.0
django-modeltranslation==0.19.9
django-phonenumber-field==8.0.0
django-prometheus==2.3.1
django-simple-history==3.7.0
django-timezone-field==7.0
django-unfold==0.38.0
djangorestframework==3.15.2
drf-yasg==1.21.7
frozenlist==1.4.1
idna==3.8
inflection==0.5.1
Jinja2==3.1.4
kombu==5.4.0
magic-filter==1.0.12
Markdown==3.7
MarkupSafe==2.1.5
multidict==6.1.0
packaging==24.1
phonenumbers==8.13.45
prometheus_client==0.20.0
prompt_toolkit==3.0.47
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tablib==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
yarl==1.11.1
gunicorn==23.0.0
redis==5.0.8
psycopg2==2.9.9